requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.1
    # via streamlit
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   streamlit
    #   uvicorn
coloredlogs==15.0.1
    # via
    #   onnxruntime
    #   onnxruntime-gpu
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.8
    # via -r requirements.in
filelock==3.17.0
    # via torch
flatbuffers==25.1.24
    # via
    #   onnxruntime
    #   onnxruntime-gpu
fonttools==4.55.8
    # via matplotlib
fsspec==2025.2.0
    # via torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.14.0
    # via uvicorn
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
    #   torch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.0
    # via
    #   seaborn
    #   ultralytics
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
narwhals==1.25.0
    # via altair
networkx==3.4.2
    # via torch
numpy==2.1.1
    # via
    #   contourpy
    #   matplotlib
    #   onnxruntime
    #   onnxruntime-gpu
    #   opencv-python
    #   pandas
    #   pydeck
    #   scipy
    #   seaborn
    #   streamlit
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
onnxruntime==1.20.1
    # via -r requirements.in
onnxruntime-gpu==1.20.1
    # via -r requirements.in
opencv-python==4.11.0.86
    # via ultralytics
packaging==24.2
    # via
    #   altair
    #   matplotlib
    #   onnxruntime
    #   onnxruntime-gpu
    #   streamlit
pandas==2.2.3
    # via
    #   seaborn
    #   streamlit
    #   ultralytics
pillow==11.1.0
    # via
    #   matplotlib
    #   streamlit
    #   torchvision
    #   ultralytics
protobuf==5.29.3
    # via
    #   onnxruntime
    #   onnxruntime-gpu
    #   streamlit
psutil==6.1.1
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pyarrow==19.0.0
    # via streamlit
pydantic==2.10.6
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via -r requirements.in
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via ultralytics
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   streamlit
    #   ultralytics
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
scipy==1.15.1
    # via
    #   -r requirements.in
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
starlette==0.45.3
    # via fastapi
streamlit==1.42.0
    # via -r requirements.in
sympy==1.13.1
    # via
    #   onnxruntime
    #   onnxruntime-gpu
    #   torch
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
torch==2.6.0
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.21.0
    # via ultralytics
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via ultralytics
triton==3.2.0
    # via torch
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   referencing
    #   rich
    #   streamlit
    #   torch
    #   uvicorn
tzdata==2025.1
    # via pandas
ultralytics==8.3.71
    # via -r requirements.in
ultralytics-thop==2.0.14
    # via ultralytics
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
watchdog==6.0.0
    # via streamlit
